Separate E2EE witness test suites
